requirements.txt
python-telegram-bot==21.4
openai>=1.42.0
python-dotenv
